requirements.txt
django==2.2
djangorestframework==3.9.2
uWSGI>=2.0.18,<2.1
